Gentry forest transect dataset as a phyloseq object

A dataset containing species abundance data, phylogenetic tree, sample metadata, and taxonomy information for forest plots from the Gentry dataset. This dataset has been processed to aggregate species counts by unique trinomial names, mapped onto a phylogenetic tree using the S.PhyloMaker tool, and organized into a phyloseq-class object.

Usage

gentry

Format

A phyloseq-class object containing:

OTU table

Species abundance counts, with rows as samples (sites) and columns as species identified by their trinomial names.

phylogenetic tree

A phylo object representing the evolutionary relationships among species in the dataset.

sample data

Metadata for each forest plot site, including geographic coordinates and other site attributes.

taxonomy table

Taxonomic classifications (family, genus, species) matched using the taxonlookup package.

Source

Raw data originally from: https://github.com/zdealveindy/anadat-r/tree/master/data

Phylogenetic tree and nodes from: https://github.com/jinyizju/S.PhyloMaker

Taxonomic lookups performed using the taxonlookup package.

Details

This dataset was constructed by combining raw species presence data from the Gentry forest plot dataset with phylogenetic information and cleaned taxonomic names. The species names were standardized by removing uncertain annotations such as "cf." and "aff." and mapped to a phylogenetic tree using S.PhyloMaker with scenario S1. Species not matched to the phylogeny (less than 2% of counts) were removed from the final dataset.

The data can be used for analyses involving phylogenetic diversity, community ecology, and evolutionary relationships among species in forest plots.

Examples

library(phyloseq)
data(gentry)

# view the OTU table (species counts)
otu_table(gentry)[1:5, 1:20]
#> OTU Table:          [20 taxa and 5 samples]
#>                      taxa are columns
#>          Abies_lasiocarpa Abies_amabilis Picea_sitchensis Picea_abies
#> ankarif                 0              0                0           0
#> banyong                 0              0                0           0
#> beza1                   0              0                0           0
#> beza2                   0              0                0           0
#> brisefer                0              0                0           0
#>          Pinus_sylvestris Fitzroya_cupressoides Juniperus_communis
#> ankarif                 0                     0                  0
#> banyong                 0                     0                  0
#> beza1                   0                     0                  0
#> beza2                   0                     0                  0
#> brisefer                0                     0                  0
#>          Juniperus_virginiana Agathis_microstachys Allocasia_macrorhiza
#> ankarif                     0                    0                    0
#> banyong                     0                    0                    0
#> beza1                       0                    0                    0
#> beza2                       0                    0                    0
#> brisefer                    0                    0                    0
#>          Rhodopspatha_m1 Rhodospatha_m1 Rhodospatha_latifolia
#> ankarif                0              0                     0
#> banyong                0              0                     0
#> beza1                  0              0                     0
#> beza2                  0              0                     0
#> brisefer               0              0                     0
#>          Symplocarpus_purpusii Anthurium_weberbaueri Anthurium_giganteum
#> ankarif                      0                     0                   0
#> banyong                      0                     0                   0
#> beza1                        0                     0                   0
#> beza2                        0                     0                   0
#> brisefer                     0                     0                   0
#>          Anthurium_formosum Anthurium_dolichostachyum Anthurium_cordiforme
#> ankarif                   0                         0                    0
#> banyong                   0                         0                    0
#> beza1                     0                         0                    0
#> beza2                     0                         0                    0
#> brisefer                  0                         0                    0
#>          Anthurium_argyrostachyum
#> ankarif                         0
#> banyong                         0
#> beza1                           0
#> beza2                           0
#> brisefer                        0

# the phylogenetic tree
phy_tree(gentry)
#> 
#> Phylogenetic tree with 585 tips and 349 internal nodes.
#> 
#> Tip labels:
#>   Abies_lasiocarpa, Abies_amabilis, Picea_sitchensis, Picea_abies, Pinus_sylvestris, Fitzroya_cupressoides, ...
#> Node labels:
#>   Spermatophyta, , Pinales, , , , ...
#> 
#> Rooted; includes branch length(s).

# view the sample data
sample_data(gentry) |> head()
#>             Lat  Long Elev Precip
#> ankarif  -16.32 46.82   85   1400
#> banyong    5.00  9.17  420   2600
#> beza1    -23.63 44.57  150    600
#> beza2    -23.63 44.57  175    600
#> brisefer -20.38 57.43  550   2400
#> korup      5.73  8.70  100   5460

# extract taxonomic classifications
tax_table(gentry) |> head()
#> Taxonomy Table:     [6 taxa by 4 taxonomic ranks]:
#>                       genus      family         order     group        
#> Abies_lasiocarpa      "Abies"    "Pinaceae"     "Pinales" "Gymnosperms"
#> Abies_amabilis        "Abies"    "Pinaceae"     "Pinales" "Gymnosperms"
#> Picea_sitchensis      "Picea"    "Pinaceae"     "Pinales" "Gymnosperms"
#> Picea_abies           "Picea"    "Pinaceae"     "Pinales" "Gymnosperms"
#> Pinus_sylvestris      "Pinus"    "Pinaceae"     "Pinales" "Gymnosperms"
#> Fitzroya_cupressoides "Fitzroya" "Cupressaceae" "Pinales" "Gymnosperms"