Compute normalized eigenvalues for the MPQr distance

Given an object containing the eigendecomposition of the matrix \(\T\) used for the PQ distance, , this function computes the vector of normalized weights corresponding to the parameter \(r \in [0,1]\).

Usage

get_dr(eig, r)

Arguments

eig

An 'eigen' object. Should be obtained by calling 'eigen' on the covariance matrix obtained from a phylogenetic tree T.

r

Numeric scalar in \([0,1]\) specifying the tuning parameter of the \(\MPQr\) distance.

Value

A numeric vector of length length(eig$values).

Examples

eig <- eigen(matrix(c(2,1,1,2), 2, 2))
# when r = 0, the eigenvalues are equal
mpqDist:::get_dr(eig, r = 0)
#> [1] 0.5 0.5
mpqDist:::get_dr(eig, r = 0.5)
#> [1] 0.6428571 0.3571429
# when r = 1, the eigenvalues are proportional to the original eigenvalues
mpqDist:::get_dr(eig, r = 1)
#> [1] 0.75 0.25