Subsampled OTU Table from the Gentry Dataset

A small subset of the OTU (species abundance) table from the gentry phyloseq object, containing counts for 50 randomly selected taxa (out of the original 585). This object is primarily intended for use in examples and demonstrations, and is not recommended for substantive ecological or phylogenetic analyses.

Usage

small_otutab

Format

A numeric matrix with 50 columns (taxa) and rows corresponding to samples/sites.

Details

The subset was created by randomly sampling 50 taxa indices and subsetting the OTU table from the gentry dataset using subset_taxa(). The resulting OTU table was coerced to a matrix for convenience.

See also

Examples

dim(small_otutab)
#> [1] 197  50
## note that because we subsampled taxa, many of the sites are exactly the same and have no taxa
head(small_otutab)
#> OTU Table:          [50 taxa and 6 samples]
#>                      taxa are columns
#>          Philodendron_consanguineum Bactris_m1 Aiphanes_aculeata
#> ankarif                           0          0                 0
#> banyong                           0          0                 0
#> beza1                             0          0                 0
#> beza2                             0          0                 0
#> brisefer                          0          0                 0
#> korup                             0          0                 0
#>          Cyathostemma_m1 Bocageopsis?_m1 Malmea_dielsiana Oxandra_m1
#> ankarif                0               0                0          0
#> banyong                0               0                0          0
#> beza1                  0               0                0          0
#> beza2                  0               0                0          0
#> brisefer               0               0                0          0
#> korup                  0               0                0          0
#>          Oxandra_venezuelana Cremastosperma_gracilipes Mitrephora_m1
#> ankarif                    0                         0             0
#> banyong                    0                         0             0
#> beza1                      0                         0             0
#> beza2                      0                         0             0
#> brisefer                   0                         0             0
#> korup                      0                         0             0
#>          Sapranthus_foetidus Fissistigma_glaucescens Xylopia_bocatorena
#> ankarif                    0                       0                  0
#> banyong                    0                       0                  0
#> beza1                      0                       0                  0
#> beza2                      0                       0                  0
#> brisefer                   0                       0                  0
#> korup                      0                       0                  0
#>          Guatteria_m4 Mahonia_leschnaultii Bursera_simarouba Bursera_simaruba
#> ankarif             0                    0                 0                0
#> banyong             0                    0                 0                0
#> beza1               0                    0                 0                0
#> beza2               0                    0                 0                0
#> brisefer            0                    0                 0                0
#> korup               0                    0                 0                0
#>          Parishia_minor Pseudospondias_microcarpa Operculicarya_decaryi
#> ankarif               0                         0                     0
#> banyong               0                         1                     0
#> beza1                 0                         0                     0
#> beza2                 0                         0                     4
#> brisefer              0                         0                     0
#> korup                 0                         0                     0
#>          Poupartia_minor Semecarpus_m1 Trichoscypha_m2 Mangifera_m1
#> ankarif                0             0               0            0
#> banyong                0             0               0            0
#> beza1                  0             0               0            0
#> beza2                 11             0               0            0
#> brisefer               0             0               0            0
#> korup                  0             0               0            0
#>          Mangifera_indica Protorhus_m1 Protorhus_m3 Astronium_m1 Acer_caesium
#> ankarif                 0            0            0            0            0
#> banyong                 0            0            0            0            0
#> beza1                   0            0            0            0            0
#> beza2                   0            0            0            0            0
#> brisefer                0            0            0            0            0
#> korup                   0            0            0            0            0
#>          Weinmannia_trichosperma Ostrya_virginiana Betula_papyrifera
#> ankarif                        0                 0                 0
#> banyong                        0                 0                 0
#> beza1                          0                 0                 0
#> beza2                          0                 0                 0
#> brisefer                       0                 0                 0
#> korup                          0                 0                 0
#>          Bauhinia_malabarica Iresine_iterrupta Ilex_m4 Ilex_m3
#> ankarif                    0                 0       0       0
#> banyong                    0                 0       0       0
#> beza1                      0                 0       0       0
#> beza2                      0                 0       0       0
#> brisefer                   0                 0       0       0
#> korup                      0                 0       0       0
#>          Schefflera_morototoni Polyscias_australiana Liabum_caducifolium
#> ankarif                      0                     0                   0
#> banyong                      0                     0                   0
#> beza1                        0                     0                   0
#> beza2                        0                     0                   0
#> brisefer                     0                     0                   0
#> korup                        0                     0                   0
#>          Aspidosperma_curranii Aspidosperma_m1 Aspidosperma_pyrifolium
#> ankarif                      0               0                       0
#> banyong                      0               0                       0
#> beza1                        0               0                       0
#> beza2                        0               0                       0
#> brisefer                     0               0                       0
#> korup                        0               0                       0
#>          Alstonia_comptonii Prestonia_obovata Peltastes_m1
#> ankarif                   0                 0            0
#> banyong                   0                 0            0
#> beza1                     0                 0            0
#> beza2                     0                 0            0
#> brisefer                  0                 0            0
#> korup                     0                 0            0
#>          Habracanthus_sanguineus Mendoncia_retusa Justicia_aurea
#> ankarif                        0                0              0
#> banyong                        0                0              0
#> beza1                          0                0              0
#> beza2                          0                0              0
#> brisefer                       0                0              0
#> korup                          0                0              0
#>          Ceratophytum_tetragonolobum Tabebuia_heterophylla
#> ankarif                            0                     0
#> banyong                            0                     0
#> beza1                              0                     0
#> beza2                              0                     0
#> brisefer                           0                     0
#> korup                              0                     0